# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ini adalah dokumentasi API untuk platform OtoMeet.",
        "title": "OtoMeet API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/diskusi/latest": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil semua pesan terbaru di atas ID tertentu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diskusi"
                ],
                "summary": "Ambil Pesan Diskusi Terbaru Global",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Last Message ID",
                        "name": "last_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Diskusi"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            }
        },
        "/api/diskusi/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui teks pesan diskusi yang sudah ada. Hanya pemilik pesan yang bisa mengeditnya.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diskusi"
                ],
                "summary": "Edit Pesan Diskusi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Diskusi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Teks Pesan Baru",
                        "name": "diskusi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DiskusiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger403Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus pesan diskusi. Hanya pemilik pesan yang bisa menghapusnya.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diskusi"
                ],
                "summary": "Hapus Pesan Diskusi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Diskusi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger403Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            }
        },
...
        "/api/touring/{id}/diskusi": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil semua komentar/diskusi untuk suatu agenda touring",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diskusi"
                ],
                "summary": "Ambil Diskusi Touring",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Touring ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Diskusi"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan komentar baru ke dalam agenda touring tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diskusi"
                ],
                "summary": "Kirim Komentar Diskusi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Touring ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pesan Diskusi",
                        "name": "diskusi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DiskusiRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.DiskusiRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Halo semuanya, kapan kita kumpul?"
                }
            }
        },
...
        "model.Diskusi": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_edited": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "touring": {
                    "$ref": "#/definitions/model.Touring"
                },
                "touring_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Touring": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deskripsi": {
                    "type": "string"
                },
                "harga_tiket": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "kuota": {
                    "type": "integer"
                },
                "lokasi_akhir": {
                    "type": "string"
                },
                "lokasi_awal": {
                    "type": "string"
                },
                "nama_touring": {
                    "type": "string"
                },
                "tanggal": {
                    "type": "string"
                },
                "tujuan": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "waktu": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "description": "Changed: removed NOT NULL to handle existing data",
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "profile_photo": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "vehicles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Vehicle"
                    }
                }
            }
        },
        "model.Vehicle": {
            "type": "object",
            "properties": {
                "bike_model": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_plate": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "utils.Swagger401Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "token tidak valid atau sudah expired"
                }
            }
        },
        "utils.Swagger403Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "user tidak memiliki akses untuk fitur ini"
                }
            }
        },
        "utils.SwaggerBasicResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "string"
                },
                "error": {
                    "type": "string",
                    "example": "detail error"
                },
                "message": {
                    "type": "string",
                    "example": "detail pesan"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}